Rename patient DetailView files to match PatientView
