+++ client/packages/system/src/Patient/PatientView/index.ts
@@ -1,4 +1,4 @@
-export { PatientView } from './DetailView';
+export { PatientView } from './PatientView';
 
 export enum PatientModal {
   Prescription = 'PRESCRIPTION',
